browser/devtools/webide/test/test_duplicate_import.html

<!DOCTYPE html>

<html>

  <head>
    <meta charset="utf8">
    <title></title>

    <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
    <script type="application/javascript" src="chrome://mochikit/content/chrome-harness.js"></script>
    <script type="application/javascript;version=1.8" src="head.js"></script>
    <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  </head>

  <body>

    <script type="application/javascript;version=1.8">
      window.onload = function() {
        SimpleTest.waitForExplicitFinish();

        Task.spawn(function* () {
          Services.prefs.setBoolPref("devtools.webide.sidebars", false);
          let win = yield openWebIDE();
          let packagedAppLocation = getTestFilePath("app");
          let hostedAppManifest = TEST_BASE + "hosted_app.manifest";

          yield win.AppProjects.load();
          is(win.AppProjects.store.object.projects.length, 0, "IDB is empty");

          info("to call importPackagedApp(" + packagedAppLocation + ")");
          yield win.projectList.importPackagedApp(packagedAppLocation);
          yield waitForUpdate(win, "project-validated");
          yield nextTick();

          info("to call importHostedApp(" + hostedAppManifest + ")");
          yield win.projectList.importHostedApp(hostedAppManifest);
          yield waitForUpdate(win, "project-validated");
          yield nextTick();

          info("to call importPackagedApp(" + packagedAppLocation + ") again");
          yield win.projectList.importPackagedApp(packagedAppLocation);
          yield waitForUpdate(win, "project-validated");

          let project = win.AppManager.selectedProject;
          is(project.location, packagedAppLocation, "Correctly reselected existing packaged app.");
          yield nextTick();

          info("to call importHostedApp(" + hostedAppManifest + ") again");
          yield win.projectList.importHostedApp(hostedAppManifest);
          yield waitForUpdate(win, "project-validated");
          project = win.AppManager.selectedProject;
          is(project.location, hostedAppManifest, "Correctly reselected existing hosted app.");
          yield nextTick();

          info("opening panel");
          yield win.Cmds.showProjectPanel();
          info("panel open");

          let panelNode = win.document.querySelector("#project-panel");
          let items = panelNode.querySelectorAll(".panel-item");
          // 3 controls, + 2 projects
          is(items.length, 5, "5 projects in panel");
          is(items[3].getAttribute("label"), "A name (in app directory)", "Panel label is correct");
          is(items[4].getAttribute("label"), "hosted manifest name property", "Panel label is correct");

          yield closeWebIDE(win);

          yield removeAllProjects();

          SimpleTest.finish();
        }).then(null, e => {
          ok(false, "Exception: " + e);
          SimpleTest.finish();
        });
      }
    </script>
  </body>
</html>
